refactor: rename handlers to handles and update related classes and tests
